+++ hs_core/urls.py
@@ -44,7 +44,7 @@ urlpatterns = [
         views.unshare_resource_with_group, name='unshare_resource_with_group'),
     url(r'^_internal/(?P<shortkey>[0-9a-f-]+)/undo-share-resource-with-group/(?P<group_id>[0-9]+)/$',
         views.undo_share_resource_with_group, name='undo_share_resource_with_group'),
-    url(r'^_internal/create-user-group/$', views.create_user_group, name='create_user_group'),
+    url(r'^_internal/create-user-group/$', views.group.create_user_group, name='create_user_group'),
     url(r'^_internal/update-user-group/(?P<group_id>[0-9]+)$', views.update_user_group,
         name='update_user_group'),
     url(r'^_internal/delete-user-group/(?P<group_id>[0-9]+)$', views.delete_user_group,
